Fixed tests to pass in Python 3.
The simmetrica code uses a dictionary and range in a few places. This
causes some small Python 3 compatability issues.

The order of the dictionary isn't guaranteed, but the test expects it
to be. This fix changes that so it works in Python 3 by sorting the
result before comparing.

In Python 3 range() doesn't return a list, but rather returns a range
object. Therefore, to make this work in both versions the test makes
sure its a list before comparing.